Added ability to connect to PostgreSQL based on an environment variable (ASTROFEED_POSTGRES) which default to False so old behavior is fallback. Requires addition of psycopg2-binary dependency for PostgreSQL driver.
Added bootstrapping SQL files to create the database structure in Postgres, including the tables, indexes, sequences, and materialized view for the feed statistics. This can be run by executing plsql with an input of "bootstrap.sql", which will call all the other necessary SQL files. No data is as-of-yet included in these files.